Change localization to use mediapipe
