--- Assets/Scripts/PythonScripts/haarcascade_frontalface_default.xml.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 271d4f3771648884482c50411185e247
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
