+++ Assets/Scripts/PythonScripts/localization_mediapipe.py.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b94974e9e3d730541ac9c25099927287
+guid: bf22719c60c1a2d4cb60c6f0b00133a6
 DefaultImporter:
   externalObjects: {}
   userData:
